# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Clientes API
  version: 1.0.0
  description: API para gerenciamento de clientes, produtos e estágios
servers:
  - url: http://localhost:3000
    description: Servidor local
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Address:
      type: object
      required:
        - postalCode
      properties:
        postalCode:
          type: string
          description: CEP do endereço
        street:
          type: string
        complement:
          type: string
        unit:
          type: string
        district:
          type: string
        city:
          type: string
        stateCode:
          type: string
        state:
          type: string
        region:
          type: string
    Product:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        value:
          type: number
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    StageHistory:
      type: object
      properties:
        from:
          type: string
        to:
          type: string
        at:
          type: string
          format: date-time
        by:
          type: string
        note:
          type: string
    Customer:
      type: object
      properties:
        _id:
          type: string
        document:
          type: string
        person:
          type: string
          enum: [F, J]
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        active:
          type: boolean
        address:
          $ref: '#/components/schemas/Address'
        stage:
          type: string
        stageChangedAt:
          type: string
          format: date-time
        stageHistory:
          type: array
          items:
            $ref: '#/components/schemas/StageHistory'
        products:
          type: array
          items:
            $ref: '#/components/schemas/Product'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateCustomer:
      type: object
      required:
        - document
        - person
        - name
        - email
        - phone
        - address
      properties:
        document:
          type: string
          description: CPF ou CNPJ do cliente
        person:
          type: string
          enum: [F, J]
          description: F para Pessoa Física ou J para Pessoa Jurídica
        name:
          type: string
          description: Nome do cliente
        email:
          type: string
          format: email
          description: Email do cliente
        phone:
          type: string
          description: Telefone do cliente
        address:
          type: object
          required:
            - postalCode
          properties:
            postalCode:
              type: string
              description: CEP do endereço
            street:
              type: string
            complement:
              type: string
            unit:
              type: string
            district:
              type: string
            city:
              type: string
            stateCode:
              type: string
            state:
              type: string
            region:
              type: string
...
    Error:
      type: object
      properties:
        message:
          type: string
...
paths:
...
  /customers:
    get:
      tags:
        - Customers
      summary: Listar todos os clientes ativos
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de clientes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Customer'
    
    post:
      tags:
        - Customers
      summary: Criar novo cliente
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCustomer'
            example:
              document: "12345678900"
              person: "F"
              name: "João da Silva"
              email: "joao@example.com"
              phone: "11987654321"
              address:
                postalCode: "01310100"
      responses:
        '201':
          description: Cliente criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '400':
          description: Dados inválidos ou cliente já existe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...